Move forEachState function to its own file
